+++ icat/helper.py
@@ -34,7 +34,9 @@ True
 """
 
 import sys
+from contextlib import contextmanager
 import datetime
+import logging
 import suds.sax.date
 
 
@@ -219,3 +221,14 @@ def ms_timestamp(dt):
             dt = dt.replace(tzinfo=None) - offs
         ts = 1000 * (dt - datetime.datetime(1970, 1, 1)).total_seconds()
     return int(ts)
+
+
+@contextmanager
+def disable_logger(name):
+    """Context manager to temporarily disable a logger.
+    """
+    logger = logging.getLogger(name)
+    sav_state = logger.disabled
+    logger.disabled = True
+    yield
+    logger.disabled = sav_state